refactor(minesweeper): 패키지 분리
