+++ src/test/kotlin/minesweeper/MineBoardTest.kt
@@ -1,5 +1,6 @@
 package minesweeper
 
+import minesweeper.domain.MineBoard
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
